+++ src/main/java/com/ngtesting/platform/service/impl/CustomFieldServiceImpl.java
@@ -61,7 +61,7 @@ public class CustomFieldServiceImpl extends BaseServiceImpl implements CustomFie
                 customFieldOptionDao.saveAll(vo.getId(), vo.getOptions());
             }
         } else {
-            if (vo.getOrgId() != orgId) {
+            if (vo.getOrgId().intValue() != orgId.intValue()) {
                 return null;
             }
